Refactor file transfer logic into separate handler
