+++ worker.py
@@ -8,10 +8,6 @@ import logging
 import tkinter
 import queue
 import struct
-import os
-import shutil
-import tempfile
-import zipfile
 from typing import Optional
 import msgpack
 import pyperclip
@@ -19,6 +15,7 @@ from pynput import mouse, keyboard
 from zeroconf import ServiceInfo, Zeroconf, ServiceBrowser
 from monitorcontrol import get_monitors
 from PySide6.QtCore import QObject, Signal, QSettings
+from file_transfer import FileTransferHandler
 from config import (
     SERVICE_TYPE,
     SERVICE_NAME_PREFIX,
@@ -54,7 +51,7 @@ class KVMWorker(QObject):
         'active_client', 'pynput_listeners', 'zeroconf', 'streaming_thread',
         'switch_monitor', 'local_ip', 'server_ip', 'connection_thread',
         'device_name', 'clipboard_thread', 'last_clipboard', 'server_socket',
-        'network_file_clipboard', '_cancel_transfer', 'last_server_ip'
+        'last_server_ip', 'file_handler'
     )
 
     finished = Signal()
@@ -87,9 +84,7 @@ class KVMWorker(QObject):
         self.clipboard_thread = None
         self.last_clipboard = ""
         self.server_socket = None
-        self.network_file_clipboard = None
-        logging.debug("Network file clipboard cleared")
-        self._cancel_transfer = threading.Event()
+        self.file_handler = FileTransferHandler(self)
         self._host_mouse_controller = None
         self._orig_mouse_pos = None
 
@@ -172,469 +167,18 @@ class KVMWorker(QObject):
             time.sleep(0.5)
 
     # ------------------------------------------------------------------
-    # File transfer helpers
-    # ------------------------------------------------------------------
-    def _get_temp_dir(self) -> str:
-        """
-        Creates and returns the path to a dedicated temporary directory for the app.
-        Uses the custom path from settings if available, otherwise the system default.
-        """
-        base_path = self.settings.get('temp_path') or tempfile.gettempdir()
-
-        # Construct the root path for all temporary files using the configured
-        # drive and the constant directory structure.
-        app_temp_path = os.path.join(base_path, *TEMP_DIR_PARTS)
-
-        try:
-            os.makedirs(app_temp_path, exist_ok=True)
-            transfer_temp_dir = tempfile.mkdtemp(dir=app_temp_path)
-            logging.info(f"Using temporary directory: {transfer_temp_dir}")
-            return transfer_temp_dir
-        except OSError as e:
-            logging.error(
-                f"Could not create temporary directory at {app_temp_path}: {e}. Falling back to system default."
-            )
-            return tempfile.mkdtemp()
-
-    def _create_archive(self, paths, cancel_event: Optional[threading.Event] = None):
-        temp_dir = self._get_temp_dir()
-        archive = os.path.join(temp_dir, 'share.zip')
-        logging.debug("Created temp archive dir %s", temp_dir)
-        # Log available space in the temporary directory which will hold the archive
-        try:
-            usage = shutil.disk_usage(temp_dir)
-            logging.debug(
-                "Temp dir disk usage - total: %s, used: %s, free: %s",
-                usage.total,
-                usage.used,
-                usage.free,
-            )
-        except Exception as e:
-            logging.debug("Failed to query temp dir disk usage: %s", e)
-        start_time = time.time()
-        try:
-            # Pre-scan all paths to determine total number of files
-            total_files = 0
-            for p in paths:
-                if os.path.isdir(p):
-                    for _, _, files in os.walk(p):
-                        total_files += len(files)
-                else:
-                    total_files += 1
-
-            archived_files = 0
-            with zipfile.ZipFile(archive, 'w', zipfile.ZIP_DEFLATED, allowZip64=True) as zf:
-                def _write_with_progress(src_path, arcname, ctype):
-                    file_size = os.path.getsize(src_path)
-                    if cancel_event and cancel_event.is_set():
-                        raise RuntimeError('archive canceled')
-
-                    last_percentage = -1
-                    last_emit_time = time.time()
-                    start_time = time.time()
-
-                    if file_size > 1_000_000_000:
-                        info = zipfile.ZipInfo(arcname, date_time=time.localtime(time.time())[:6])
-                        info.compress_type = ctype
-                        with open(src_path, 'rb') as src_file, zf.open(info, 'w', force_zip64=True) as dest:
-                            sent = 0
-                            while True:
-                                if cancel_event and cancel_event.is_set():
-                                    raise RuntimeError('archive canceled')
-                                chunk = src_file.read(FILE_CHUNK_SIZE)
-                                if not chunk:
-                                    break
-                                dest.write(chunk)
-                                sent += len(chunk)
-
-                                current_percentage = int((sent / file_size) * 100)
-                                if current_percentage > last_percentage or time.time() - last_emit_time > PROGRESS_UPDATE_INTERVAL:
-                                    # --- Speed and ETR Calculation ---
-                                    elapsed_time = time.time() - start_time
-                                    speed_mbps = (sent / (1024*1024)) / elapsed_time if elapsed_time > 0 else 0
-                                    remaining_bytes = file_size - sent
-                                    etr_seconds = int(remaining_bytes / (speed_mbps * 1024 * 1024)) if speed_mbps > 0 else 0
-                                    etr_str = time.strftime('%M:%S', time.gmtime(etr_seconds)) if etr_seconds < 3600 else time.strftime('%H:%M:%S', time.gmtime(etr_seconds))
-
-                                    label = f"{os.path.basename(src_path)}: {sent/1024/1024:.1f}MB / {file_size/1024/1024:.1f}MB\n"
-                                    label += f"Sebesség: {speed_mbps:.1f} MB/s | Hátralévő idő: {etr_str}"
-
-                                    self.update_progress_display.emit(current_percentage, label)
-                                    last_percentage = current_percentage
-                                    last_emit_time = time.time()
-                        final_label = f"{os.path.basename(src_path)}: Kész! ({file_size/1024/1024:.1f}MB)"
-                        self.update_progress_display.emit(100, final_label)
-                    else:
-                        zf.write(src_path, arcname, compress_type=ctype)
-                for p in paths:
-                    if os.path.isdir(p):
-                        base = os.path.basename(p.rstrip(os.sep))
-                        for root, _, files in os.walk(p):
-                            for f in files:
-                                full = os.path.join(root, f)
-                                rel = os.path.join(base, os.path.relpath(full, p))
-                                file_size = os.path.getsize(full)
-                                if file_size > 1_000_000_000:
-                                    logging.info(
-                                        "Adding large file %s (%d bytes) to archive",
-                                        full,
-                                        file_size,
-                                    )
-                                try:
-                                    compress_type = zipfile.ZIP_DEFLATED
-                                    ext = os.path.splitext(f)[1].lower()
-                                    if file_size > 1_000_000_000 or ext in {'.mkv', '.mp4', '.mov'}:
-                                        compress_type = zipfile.ZIP_STORED
-                                    _write_with_progress(full, rel, compress_type)
-                                    logging.debug("Archived %s", full)
-                                    if cancel_event and cancel_event.is_set():
-                                        raise RuntimeError('archive canceled')
-                                except MemoryError:
-                                    msg = (
-                                        f"Archiv\xe1l\xe1si hiba: Kev\xe9s a mem\xf3ria a(z) {os.path.basename(full)} t\xf6m\xf6r\xedt\xe9s\xe9hez."
-                                    )
-                                    logging.error(msg, exc_info=True)
-                                    self.file_transfer_error.emit(msg)
-                                    return None
-                                except (IOError, OSError) as e:
-                                    msg = f"Archiv\xe1l\xe1si hiba: Nincs el\xe9g hely vagy IO probl\xe9ma ({e})."
-                                    logging.error(msg, exc_info=True)
-                                    self.file_transfer_error.emit(msg)
-                                    return None
-                                except zipfile.LargeZipFile as e:
-                                    msg = f"Archiv\xe1l\xe1si hiba: {e}"
-                                    logging.error(msg, exc_info=True)
-                                    self.file_transfer_error.emit(msg)
-                                    return None
-                                archived_files += 1
-                                percentage = int(archived_files / total_files * 100) if total_files else 0
-                                label = f"Tömörítés: {os.path.basename(full)} ({archived_files}/{total_files} fájl)"
-                                self.update_progress_display.emit(percentage, label)
-                    else:
-                        file_size = os.path.getsize(p)
-                        if file_size > 1_000_000_000:
-                            logging.info(
-                                "Adding large file %s (%d bytes) to archive",
-                                p,
-                                file_size,
-                            )
-                        try:
-                            compress_type = zipfile.ZIP_DEFLATED
-                            ext = os.path.splitext(p)[1].lower()
-                            if file_size > 1_000_000_000 or ext in {'.mkv', '.mp4', '.mov'}:
-                                compress_type = zipfile.ZIP_STORED
-                            _write_with_progress(p, os.path.basename(p), compress_type)
-                            logging.debug("Archived %s", p)
-                            if cancel_event and cancel_event.is_set():
-                                raise RuntimeError('archive canceled')
-                        except MemoryError:
-                            msg = (
-                                f"Archiv\xe1l\xe1si hiba: Kev\xe9s a mem\xf3ria a(z) {os.path.basename(p)} t\xf6m\xf6r\xedt\xe9s\xe9hez."
-                            )
-                            logging.error(msg, exc_info=True)
-                            self.file_transfer_error.emit(msg)
-                            return None
-                        except (IOError, OSError) as e:
-                            msg = f"Archiv\xe1l\xe1si hiba: Nincs el\xe9g hely vagy IO probl\xe9ma ({e})."
-                            logging.error(msg, exc_info=True)
-                            self.file_transfer_error.emit(msg)
-                            return None
-                        except zipfile.LargeZipFile as e:
-                            msg = f"Archiv\xe1l\xe1si hiba: {e}"
-                            logging.error(msg, exc_info=True)
-                            self.file_transfer_error.emit(msg)
-                            return None
-                        archived_files += 1
-                        percentage = int(archived_files / total_files * 100) if total_files else 0
-                        label = f"Tömörítés: {os.path.basename(p)} ({archived_files}/{total_files} fájl)"
-                        self.update_progress_display.emit(percentage, label)
-            label = f"Tömörítés kész. ({os.path.basename(archive)})"
-            self.update_progress_display.emit(100, label)
-        except Exception as e:
-            logging.error("Failed to create archive: %s", e, exc_info=True)
-            try:
-                shutil.rmtree(temp_dir, ignore_errors=True)
-            except Exception as cleanup_err:
-                logging.error("Failed to cleanup temp dir %s: %s", temp_dir, cleanup_err)
-            self.file_transfer_error.emit(f"Archive creation failed: {e}")
-            return None
-        duration = time.time() - start_time
-        if duration > 600:
-            logging.warning("Archive creation took %.1f seconds", duration)
-        logging.debug("Archive created at %s", archive)
-        return archive
-
-    def _safe_extract_archive(self, archive_path, dest_dir):
-        """Extract archive to dest_dir and cleanup on failure."""
-        temp_extract = tempfile.mkdtemp(dir=dest_dir)
-        logging.debug("Created temporary extract dir %s", temp_extract)
-        try:
-            with zipfile.ZipFile(archive_path, 'r') as zf:
-                zf.extractall(temp_extract)
-            if self._cancel_transfer.is_set():
-                raise RuntimeError('transfer canceled')
-            for name in os.listdir(temp_extract):
-                source_path = os.path.join(temp_extract, name)
-                target_path_base = os.path.join(dest_dir, name)
-
-                # --- START NEW FILENAME CONFLICT LOGIC ---
-                final_target_path = target_path_base
-                counter = 2
-                base, ext = os.path.splitext(name)
-
-                while os.path.exists(final_target_path):
-                    if ext:
-                        new_name = f"{base} ({counter}){ext}"
-                    else:
-                        new_name = f"{name} ({counter})"
-                    final_target_path = os.path.join(dest_dir, new_name)
-                    counter += 1
-
-                if final_target_path != target_path_base:
-                    logging.info(
-                        f"Filename conflict: '{target_path_base}' exists. Renaming to '{final_target_path}'"
-                    )
-                # --- END NEW FILENAME CONFLICT LOGIC ---
-
-                shutil.move(source_path, final_target_path)
-        except Exception:
-            shutil.rmtree(temp_extract, ignore_errors=True)
-            logging.debug("Extraction failed, removed %s", temp_extract)
-            raise
-        else:
-            shutil.rmtree(temp_extract, ignore_errors=True)
-            logging.debug("Extraction complete, removed %s", temp_extract)
-
-    def _send_archive(self, sock, archive_path, dest_dir):
-        self._cancel_transfer.clear()
-        logging.debug(
-            "Entering _send_archive. cancel=%s dest=%s",
-            self._cancel_transfer.is_set(),
-            dest_dir,
-        )
-        logging.debug("Cancel flag cleared at start of _send_archive")
-        prev_to = sock.gettimeout()
-        sock.settimeout(TRANSFER_TIMEOUT)
-        try:
-            size = os.path.getsize(archive_path)
-            name = os.path.basename(archive_path)
-            meta = {
-                'type': 'file_metadata',
-                'name': name,
-                'size': size,
-                'dest': dest_dir,
-                'source_id': self.network_file_clipboard.get('source_id') if self.network_file_clipboard else self.device_name,
-            }
-            if not self._send_message(sock, meta):
-                return
-            sent = 0
-            last_percentage = -1
-            last_emit_time = time.time()
-            start_time = time.time()  # For average speed calculation
-
-            with open(archive_path, 'rb') as f:
-                while not self._cancel_transfer.is_set():
-                    chunk = f.read(FILE_CHUNK_SIZE)
-                    if not chunk:
-                        break
-                    if not self._send_message(sock, {'type': 'file_chunk', 'data': chunk}):
-                        raise IOError('send failed')
-                    sent += len(chunk)
-
-                    # Throttle the UI update
-                    if time.time() - last_emit_time >= PROGRESS_UPDATE_INTERVAL:
-                        current_percentage = int((sent / size) * 100) if size > 0 else 0
-
-                        # --- Speed and ETR Calculation ---
-                        elapsed_time = time.time() - start_time
-                        speed_mbps = (sent / (1024*1024)) / elapsed_time if elapsed_time > 0 else 0
-                        remaining_bytes = size - sent
-                        etr_seconds = int(remaining_bytes / (speed_mbps * 1024 * 1024)) if speed_mbps > 0 else 0
-                        etr_str = time.strftime('%M:%S', time.gmtime(etr_seconds)) if etr_seconds < 3600 else time.strftime('%H:%M:%S', time.gmtime(etr_seconds))
-
-                        label = f"{name}: {sent/1024/1024:.1f}MB / {size/1024/1024:.1f}MB\n"
-                        label += f"Sebesség: {speed_mbps:.1f} MB/s | Hátralévő idő: {etr_str}"
-
-                        self.update_progress_display.emit(current_percentage, label)
-                        last_emit_time = time.time()
-            if self._cancel_transfer.is_set():
-                self._send_message(sock, {'type': 'transfer_canceled'})
-                return
-            final_label = f"{name}: Kész! ({size/1024/1024:.1f}MB)"
-            self.update_progress_display.emit(100, final_label)
-            self._send_message(sock, {'type': 'file_end'})
-            logging.debug(
-                "WORKER EMITTING update_progress_display: %s %d/%d",
-                name,
-                size,
-                size,
-            )
-            logging.debug("_send_archive loop completed. cancel=%s", self._cancel_transfer.is_set())
-        except Exception as e:
-            logging.error('Error sending archive: %s', e, exc_info=True)
-            self.file_transfer_error.emit(str(e))
-        finally:
-            sock.settimeout(prev_to)
-            self._cancel_transfer.clear()
-            logging.debug("Archive send finished, cancel flag cleared")
-            logging.debug("Exiting _send_archive")
-
-    def _clear_network_file_clipboard(self):
-        """Remove any stored temporary archive and clear the clipboard info."""
-        if self.network_file_clipboard and self.network_file_clipboard.get('archive'):
-            try:
-                os.remove(self.network_file_clipboard['archive'])
-                logging.debug(
-                    "Removed temporary archive %s", self.network_file_clipboard['archive']
-                )
-            except FileNotFoundError:
-                pass
-            except Exception as e:
-                logging.error(
-                    "Failed to remove temporary archive %s: %s",
-                    self.network_file_clipboard['archive'],
-                    e,
-                )
-        self.network_file_clipboard = None
-        logging.debug("Network file clipboard cleared")
-
-    def cancel_file_transfer(self):
-        """Signal ongoing file transfer loops to cancel."""
-        self._cancel_transfer.set()
-        logging.debug("File transfer cancel signal set")
-
-    # ------------------------------------------------------------------
-    # Public API used by the GUI
+    # File transfer delegation
     # ------------------------------------------------------------------
     def share_files(self, paths, operation='copy') -> None:
-        threading.Thread(target=self._share_files_thread, args=(paths, operation), daemon=True).start()
-
-    def _share_files_thread(self, paths, operation):
-        self._cancel_transfer.clear()
-        logging.debug(
-            "Entering _share_files_thread. Cancel flag: %s",
-            self._cancel_transfer.is_set(),
-        )
-        logging.debug("Cancel flag cleared at start of _share_files_thread")
-        timeout = self.settings.get('archive_timeout_seconds', 900)
-        cancel_evt = threading.Event()
-        result = {}
-        temp_archive_dir = None
-
-        def run_archiving():
-            result['archive'] = self._create_archive(paths, cancel_event=cancel_evt)
-
-        arch_thread = threading.Thread(target=run_archiving, daemon=True)
-        arch_thread.start()
-        arch_thread.join(timeout)
-        if arch_thread.is_alive():
-            cancel_evt.set()
-            arch_thread.join(5)
-            logging.critical(
-                "Archiving of %s timed out after %.1f minutes.",
-                paths,
-                timeout / 60,
-            )
-            self.file_transfer_error.emit("Archiv\xe1l\xe1s id\u0151t\xfall\xe9p\xe9s (t\xfal nagy f\xe1jl?)")
-            if result.get('archive'):
-                shutil.rmtree(os.path.dirname(result['archive']), ignore_errors=True)
-            return
-        archive = result.get('archive')
-        try:
-            if not archive:
-                self._clear_network_file_clipboard()
-                logging.debug("Archive creation failed, exiting _share_files_thread")
-                return
-
-            temp_archive_dir = os.path.dirname(archive)
-            if self.settings['role'] == 'ado':
-                self._clear_network_file_clipboard()
-                self.network_file_clipboard = {
-                    'paths': paths,
-                    'operation': operation,
-                    'archive': archive,
-                    'source_id': self.device_name,
-                }
-                logging.debug(
-                    "Network file clipboard set: %s", self.network_file_clipboard
-                )
-                self._broadcast_message({
-                    'type': 'network_clipboard_set',
-                    'source_id': self.device_name,
-                    'operation': operation,
-                })
-            else:  # This is the client-side sending logic
-                sock = self.server_socket
-                if not sock:
-                    logging.warning('No server connection for file share')
-                    self.file_transfer_error.emit("Nincs kapcsolat a szerverrel a küldéshez.")
-                    return
-
-                # Use the robust _send_archive method for sending.
-                # The 'dest' parameter for _send_archive is not used on the sending side,
-                # but the method expects it. We can pass an empty string.
-                self._send_archive(sock, archive, dest_dir="")
-        finally:
-            if temp_archive_dir and self.settings['role'] != 'ado':
-                logging.info(f"Client-side cleanup: Removing temporary archive directory {temp_archive_dir}")
-                shutil.rmtree(temp_archive_dir, ignore_errors=True)
-            self._cancel_transfer.clear()
-            logging.debug(
-                "Exiting _share_files_thread. Network clipboard: %s",
-                self.network_file_clipboard,
-            )
+        self.file_handler.share_files(paths, operation)
 
     def request_paste(self, dest_dir) -> None:
-        self._cancel_transfer.clear()
-        logging.debug(
-            "request_paste called. role=%s cancel=%s",
-            self.settings['role'],
-            self._cancel_transfer.is_set(),
-        )
-        logging.debug("Cancel flag cleared at start of request_paste")
-        if self.settings['role'] == 'ado':
-            if not self.network_file_clipboard or not self.network_file_clipboard.get('archive'):
-                logging.warning('No shared files to paste')
-                return
-            try:
-                archive_name = os.path.basename(self.network_file_clipboard['archive']) if self.network_file_clipboard and self.network_file_clipboard.get('archive') else "archívum"
-                logging.info("[WORKER_DEBUG] Starting server-side paste (extraction) for: %s", archive_name)
-                self.update_progress_display.emit(0, f"Kibontás: {archive_name}")
-                self._safe_extract_archive(self.network_file_clipboard['archive'], dest_dir)
-                logging.info("[WORKER_DEBUG] Server-side paste (extraction) COMPLETED for: %s", archive_name)
-                self.update_progress_display.emit(100, f"{archive_name}: Feldolgozás kész!")
-            except Exception as e:
-                logging.error('Extraction failed: %s', e, exc_info=True)
-                logging.error('[WORKER_DEBUG] Server-side paste (extraction) FAILED for: %s. Error: %s', archive_name, e)
-                self.file_transfer_error.emit(f"Kibontási hiba: {e}")
-                return
-            if self.network_file_clipboard.get('operation') == 'cut':
-                src_id = self.network_file_clipboard.get('source_id')
-                if src_id == self.device_name:
-                    for pth in self.network_file_clipboard.get('paths', []):
-                        try:
-                            if os.path.isdir(pth):
-                                shutil.rmtree(pth)
-                            else:
-                                os.remove(pth)
-                        except Exception as e:
-                            logging.error('Failed to delete %s: %s', pth, e)
-                    self._clear_network_file_clipboard()
-                else:
-                    for s, name in self.client_infos.items():
-                        if name == src_id:
-                            self._send_message(s, {'type': 'delete_source', 'paths': self.network_file_clipboard.get('paths', [])})
-                            break
-                    self._clear_network_file_clipboard()
-        else:
-            sock = self.server_socket
-            if sock:
-                logging.debug("Sending paste_request to server")
-                self._send_message(sock, {'type': 'paste_request', 'destination': dest_dir})
-        logging.debug("request_paste completed. cancel=%s", self._cancel_transfer.is_set())
-        self._cancel_transfer.clear()
-        logging.debug("Cancel flag cleared at end of request_paste")
+        self.file_handler.request_paste(dest_dir)
 
+    def cancel_file_transfer(self):
+        self.file_handler.cancel_file_transfer()
+
+    # ------------------------------------------------------------------
     def set_active_client_by_name(self, name):
         """Select a connected client by name as the active target."""
         logging.debug(f"set_active_client_by_name called with name={name}")
@@ -694,9 +238,9 @@ class KVMWorker(QObject):
             self.connection_thread.join(timeout=1)
         if self.clipboard_thread and self.clipboard_thread.is_alive():
             self.clipboard_thread.join(timeout=1)
+        self.file_handler.cleanup()
         # Extra safety to avoid stuck modifier keys on exit
         self.release_hotkey_keys()
-        self._clear_network_file_clipboard()
 
     def run(self):
         logging.info(f"Worker elindítva {self.settings['role']} módban.")
@@ -877,9 +421,8 @@ class KVMWorker(QObject):
         self.client_infos[sock] = client_name
         logging.info(f"Client connected: {client_name} ({addr})")
         logging.debug(
-            "monitor_client start for %s cancel=%s",
+            "monitor_client start for %s",
             client_name,
-            self._cancel_transfer.is_set(),
         )
         # send current clipboard to newly connected client
         if self.last_clipboard:
@@ -887,16 +430,12 @@ class KVMWorker(QObject):
                 self._send_message(sock, {'type': 'clipboard_text', 'text': self.last_clipboard})
             except Exception:
                 pass
-        upload_info = None
-
         try:
             last_log = time.time()
             while self._running:
                 if time.time() - last_log >= 10:
                     logging.debug(
-                        "monitor_client main loop. cancel=%s received=%d",
-                        self._cancel_transfer.is_set(),
-                        upload_info['received'] if upload_info else 0,
+                        "monitor_client main loop.",
                     )
                     last_log = time.time()
                 try:
@@ -912,154 +451,23 @@ class KVMWorker(QObject):
                         buffer = buffer[4 + msg_len:]
                         try:
                             data = msgpack.unpackb(payload, raw=False)
-                            cmd = data.get('command')
-                            if cmd == 'switch_elitedesk':
-                                self.toggle_client_control('elitedesk', switch_monitor=True)
-                            elif cmd == 'switch_laptop':
-                                self.toggle_client_control('laptop', switch_monitor=False)
-                            elif data.get('type') == 'clipboard_text':
+                        except Exception:
+                            logging.warning("Hibas parancs a klienstol")
+                            continue
+
+                        cmd = data.get('command')
+                        if cmd == 'switch_elitedesk':
+                            self.toggle_client_control('elitedesk', switch_monitor=True)
+                        elif cmd == 'switch_laptop':
+                            self.toggle_client_control('laptop', switch_monitor=False)
+                        else:
+                            if data.get('type') == 'clipboard_text':
                                 text = data.get('text', '')
                                 if text != self.last_clipboard:
                                     self._set_clipboard(text)
                                     self._broadcast_message(data, exclude=sock)
-                            elif data.get('type') == 'paste_request':
-                                dest = data.get('destination')
-                                if self.network_file_clipboard and self.network_file_clipboard.get('archive'):
-                                    self._cancel_transfer.clear()
-                                    logging.debug("Cancel flag cleared for paste_request")
-                                    self._send_archive(sock, self.network_file_clipboard['archive'], dest)
-                            elif data.get('type') == 'file_metadata':
-                                logging.info("[WORKER_DEBUG] Received 'upload_file_start' from client: %s (size: %s)", data.get('name'), data.get('size'))
-                                temp_dir_for_download = self._get_temp_dir()
-                                incoming_path = os.path.join(temp_dir_for_download, data['name'])
-                                self._clear_network_file_clipboard()
-                                try:
-                                    incoming_file = open(incoming_path, 'wb')
-                                except Exception as e:
-                                    logging.error('Failed to open incoming file: %s', e, exc_info=True)
-                                    self.file_transfer_error.emit(str(e))
-                                    self._clear_network_file_clipboard()
-                                    break
-                                self.incoming_upload_started.emit(
-                                    data.get('name'),
-                                    data.get('size', 0)
-                                )
-                                logging.info("[WORKER_DEBUG] Emitting incoming_upload_started for: %s, size: %s", data.get('name'), data.get('size', 0))
-                                self._cancel_transfer.clear()
-                                logging.debug("Receiving upload, cancel flag cleared")
-                                sock.settimeout(TRANSFER_TIMEOUT)
-                                upload_info = {
-                                    'file': incoming_file,
-                                    'path': incoming_path,
-                                    'temp_dir': temp_dir_for_download,
-                                    'paths': data.get('paths', []),
-                                    'operation': data.get('operation', 'copy'),
-                                    'size': data.get('size', 0),
-                                    'name': data.get('name'),
-                                    'source_id': data.get('source_id', client_name),
-                                    'received': 0,
-                                    'start_time': time.time(),
-                                }
-                                last_percentage = -1
-                                last_emit_time = time.time()
-                                self.update_progress_display.emit(0, f"{upload_info['name']}: 0MB / {upload_info['size']/1024/1024:.1f}MB")
-                            elif data.get('type') == 'file_chunk':
-                                if upload_info:
-                                    try:
-                                        upload_info['file'].write(data['data'])
-                                        upload_info['received'] += len(data['data'])
-                                        if time.time() - last_emit_time >= PROGRESS_UPDATE_INTERVAL:
-                                            current_percentage = int((upload_info['received'] / upload_info['size']) * 100) if upload_info['size'] > 0 else 0
-
-                                            # --- Speed and ETR Calculation ---
-                                            elapsed_time = time.time() - upload_info['start_time']
-                                            speed_mbps = (upload_info['received'] / (1024*1024)) / elapsed_time if elapsed_time > 0 else 0
-                                            remaining_bytes = upload_info['size'] - upload_info['received']
-                                            etr_seconds = int(remaining_bytes / (speed_mbps * 1024 * 1024)) if speed_mbps > 0 else 0
-                                            etr_str = time.strftime('%M:%S', time.gmtime(etr_seconds)) if etr_seconds < 3600 else time.strftime('%H:%M:%S', time.gmtime(etr_seconds))
-
-                                            label = f"{upload_info['name']}: {upload_info['received']/1024/1024:.1f}MB / {upload_info['size']/1024/1024:.1f}MB\n"
-                                            label += f"Sebesség: {speed_mbps:.1f} MB/s | Hátralévő idő: {etr_str}"
-
-                                            self.update_progress_display.emit(current_percentage, label)
-                                            last_percentage = current_percentage
-                                            last_emit_time = time.time()
-                                        if self._cancel_transfer.is_set():
-                                            break
-                                    except Exception as e:
-                                        logging.error('Error writing chunk: %s', e, exc_info=True)
-                                        self.file_transfer_error.emit(str(e))
-                                        self._clear_network_file_clipboard()
-                                        self._cancel_transfer.set()
-                                        break
-                            elif data.get('type') == 'file_end':
-                                if upload_info:
-                                    logging.info(
-                                        "[WORKER_DEBUG] Received 'upload_file_end' for: %s",
-                                        upload_info['name'],
-                                    )
-                                    upload_info['file'].close()
-                                    final_label = f"{upload_info['name']}: Kész! ({upload_info['size']/1024/1024:.1f}MB)"
-                                    self.update_progress_display.emit(100, final_label)
-                                    self._clear_network_file_clipboard()
-                                    self.network_file_clipboard = {
-                                        'paths': upload_info['paths'],
-                                        'operation': upload_info['operation'],
-                                        'archive': upload_info['path'],
-                                        'source_id': upload_info.get('source_id', client_name),
-                                    }
-                                    logging.debug(
-                                        "Network file clipboard set: %s", self.network_file_clipboard
-                                    )
-                                    self._broadcast_message({
-                                        'type': 'network_clipboard_set',
-                                        'source_id': upload_info.get('source_id', client_name),
-                                        'operation': upload_info['operation'],
-                                    }, exclude=sock)
-                                    upload_info = None
-                                    sock.settimeout(1.0)
-                                    self._cancel_transfer.clear()
-                                    logging.debug("Upload finished, cancel flag cleared")
-                            elif data.get('type') == 'paste_success':
-                                src = data.get('source_id')
-                                if (
-                                    self.network_file_clipboard
-                                    and self.network_file_clipboard.get('operation') == 'cut'
-                                    and self.network_file_clipboard.get('source_id') == src
-                                ):
-                                    if src == self.device_name:
-                                        for pth in self.network_file_clipboard.get('paths', []):
-                                            try:
-                                                if os.path.isdir(pth):
-                                                    shutil.rmtree(pth)
-                                                else:
-                                                    os.remove(pth)
-                                            except Exception as e:
-                                                logging.error("Failed to delete %s: %s", pth, e)
-                                        self._clear_network_file_clipboard()
-                                    else:
-                                        for s2, n2 in self.client_infos.items():
-                                            if n2 == src:
-                                                self._send_message(s2, {
-                                                    'type': 'delete_source',
-                                                    'paths': self.network_file_clipboard.get('paths', []),
-                                                })
-                                                break
-                                        self._clear_network_file_clipboard()
-                            if self._cancel_transfer.is_set():
-                                if upload_info:
-                                    try:
-                                        upload_info['file'].close()
-                                        os.remove(upload_info['path'])
-                                    except Exception:
-                                        pass
-                                    upload_info = None
-                                self._clear_network_file_clipboard()
-                                sock.settimeout(1.0)
-                                self._cancel_transfer.clear()
-                                logging.debug("Upload canceled or finished, cancel flag cleared")
-                        except Exception:
-                            logging.warning("Hibas parancs a klienstol")
+                            else:
+                                self.file_handler.handle_network_message(data, sock)
                 except socket.timeout:
                     continue
                 except (socket.error, BrokenPipeError):
@@ -1070,16 +478,7 @@ class KVMWorker(QObject):
                 sock.close()
             except Exception:
                 pass
-            if upload_info and upload_info.get('temp_dir'):
-                logging.warning("Cleaning up incomplete download directory: %s", upload_info['temp_dir'])
-                try:
-                    upload_info['file'].close()
-                except Exception:
-                    pass
-                shutil.rmtree(upload_info['temp_dir'], ignore_errors=True)
-            upload_info = None
-            self._cancel_transfer.clear()
-            self._clear_network_file_clipboard()
+            self.file_handler.on_client_disconnected(sock)
             if sock in self.client_sockets:
                 self.client_sockets.remove(sock)
             if sock in self.client_infos:
@@ -1563,7 +962,7 @@ class KVMWorker(QObject):
             logging.debug(
                 "connect_to_server attempting to connect to %s cancel=%s",
                 ip,
-                self._cancel_transfer.is_set(),
+                self.file_handler._cancel_transfer.is_set(),
             )
 
             try:
@@ -1577,10 +976,8 @@ class KVMWorker(QObject):
                     settings_store = QSettings(ORG_NAME, APP_NAME)
                     settings_store.setValue('network/last_server_ip', ip)
                     self.last_server_ip = ip
-                    incoming_info = None
-                    self._cancel_transfer.clear()
-                    logging.debug("Connected to server, cancel flag cleared")
-                    logging.debug("Cancel flag state at connect: %s", self._cancel_transfer.is_set())
+                    self.file_handler._cancel_transfer.clear()
+                    logging.debug("Connected to server")
 
                     try:
                         hello = msgpack.packb({'device_name': self.device_name}, use_bin_type=True)
@@ -1661,9 +1058,8 @@ class KVMWorker(QObject):
 
                     while self._running and self.server_ip == ip:
                         logging.debug(
-                            "connect_to_server recv loop. cancel=%s received=%d",
-                            self._cancel_transfer.is_set(),
-                            incoming_info['received'] if incoming_info else 0,
+                            "connect_to_server recv loop. cancel=%s",
+                            self.file_handler._cancel_transfer.is_set(),
                         )
                         raw_len = recv_all(s, 4)
                         if not raw_len:
@@ -1705,95 +1101,8 @@ class KVMWorker(QObject):
                                 text = data.get('text', '')
                                 if text != self.last_clipboard:
                                     self._set_clipboard(text)
-                            elif event_type == 'file_metadata':
-                                temp_dir_for_download = self._get_temp_dir()
-                                incoming_tmp = os.path.join(temp_dir_for_download, data['name'])
-                                self._cancel_transfer.clear()
-                                logging.debug("Receiving file, cancel flag cleared")
-                                try:
-                                    incoming_file = open(incoming_tmp, 'wb')
-                                except Exception as e:
-                                    logging.error('Failed to open receive file: %s', e, exc_info=True)
-                                    self.file_transfer_error.emit(str(e))
-                                    break
-                                s.settimeout(TRANSFER_TIMEOUT)
-                                incoming_info = {
-                                    'path': incoming_tmp,
-                                    'dest': data['dest'],
-                                    'size': data['size'],
-                                    'name': data['name'],
-                                    'file': incoming_file,
-                                    'received': 0,
-                                    'source_id': data.get('source_id', self.device_name),
-                                    'temp_dir': temp_dir_for_download,
-                                    'start_time': time.time(),
-                                }
-                                last_percentage = -1
-                                last_emit_time = time.time()
-                                self.update_progress_display.emit(0, f"{incoming_info['name']}: 0MB / {incoming_info['size']/1024/1024:.1f}MB")
-                            elif event_type == 'file_chunk':
-                                if incoming_info:
-                                    try:
-                                        incoming_info['file'].write(data['data'])
-                                        incoming_info['received'] += len(data['data'])
-                                        current_percentage = int((incoming_info['received'] / incoming_info['size']) * 100) if incoming_info['size'] > 0 else 0
-                                        if current_percentage > last_percentage or time.time() - last_emit_time > PROGRESS_UPDATE_INTERVAL:
-                                            elapsed_time = time.time() - incoming_info['start_time']
-                                            speed_mbps = (incoming_info['received'] / (1024*1024)) / elapsed_time if elapsed_time > 0 else 0
-                                            remaining_bytes = incoming_info['size'] - incoming_info['received']
-                                            etr_seconds = int(remaining_bytes / (speed_mbps * 1024 * 1024)) if speed_mbps > 0 else 0
-                                            etr_str = time.strftime('%M:%S', time.gmtime(etr_seconds)) if etr_seconds < 3600 else time.strftime('%H:%M:%S', time.gmtime(etr_seconds))
-
-                                            label = f"{incoming_info['name']}: {incoming_info['received']/1024/1024:.1f}MB / {incoming_info['size']/1024/1024:.1f}MB\n"
-                                            label += f"Sebesség: {speed_mbps:.1f} MB/s | Hátralévő idő: {etr_str}"
-                                            self.update_progress_display.emit(current_percentage, label)
-                                            last_percentage = current_percentage
-                                            last_emit_time = time.time()
-                                        if self._cancel_transfer.is_set():
-                                            break
-                                    except Exception as e:
-                                        logging.error('Receive error: %s', e, exc_info=True)
-                                        self.file_transfer_error.emit(str(e))
-                                        self._cancel_transfer.set()
-                                        break
-                            elif event_type == 'file_end':
-                                if incoming_info:
-                                    incoming_info['file'].close()
-                                    try:
-                                        self._safe_extract_archive(incoming_info['path'], incoming_info['dest'])
-                                    finally:
-                                        shutil.rmtree(incoming_info['temp_dir'], ignore_errors=True)
-                                    self._send_message(s, {'type': 'paste_success', 'source_id': incoming_info.get('source_id')})
-                                    s.settimeout(None)
-                                    final_label = f"{incoming_info['name']}: Kész! ({incoming_info['size']/1024/1024:.1f}MB)"
-                                    self.update_progress_display.emit(100, final_label)
-                                    incoming_info = None
-                                    self._cancel_transfer.clear()
-                                    logging.debug("Download finished, cancel flag cleared")
-                            elif event_type == 'delete_source':
-                                for pth in data.get('paths', []):
-                                    try:
-                                        if os.path.isdir(pth):
-                                            shutil.rmtree(pth)
-                                        else:
-                                            os.remove(pth)
-                                    except Exception as e:
-                                        logging.error('Failed to delete %s: %s', pth, e)
-                        except Exception:
-                            logging.warning("Hibás adatcsomag")
-
-                        if self._cancel_transfer.is_set():
-                            if incoming_info:
-                                try:
-                                    incoming_info['file'].close()
-                                except Exception:
-                                    pass
-                                if incoming_info.get('temp_dir'):
-                                    shutil.rmtree(incoming_info['temp_dir'], ignore_errors=True)
-                                incoming_info = None
-                            s.settimeout(None)
-                            self._cancel_transfer.clear()
-                            logging.debug("Download canceled or finished, cancel flag cleared")
+                            else:
+                                self.file_handler.handle_network_message(data, s)
 
             except Exception as e:
                 if self._running:
@@ -1823,18 +1132,7 @@ class KVMWorker(QObject):
                     except Exception:
                         pass
                 self.release_hotkey_keys()
-                if incoming_info and incoming_info.get('temp_dir'):
-                    logging.warning(
-                        "Cleaning up incomplete download directory: %s",
-                        incoming_info['temp_dir'],
-                    )
-                    try:
-                        incoming_info['file'].close()
-                    except Exception:
-                        pass
-                    shutil.rmtree(incoming_info['temp_dir'], ignore_errors=True)
-                incoming_info = None
-                self._cancel_transfer.clear()
+                self.file_handler.on_client_disconnected(s)
                 self.server_socket = None
                 if self._running:
                     logging.info("Újracsatlakozási kísérlet 5 másodperc múlva...")
